Introduce EventBusBuilder
This makes the EventBus mostly immutable and allows removing the locks.
Only mutation is unsubscribing of subscribers and that's done by marking
them inactive with atomics.

The builder pattern also allows subscribing before publishers have registered,
which makes it easier to deal with initialization. The Build() method verifies
that every event that has been subscribed to has a registered publisher.